Move fs helpers in their own file
