+++ packages/plugins/telemetry/src/common/output/files.ts
@@ -2,7 +2,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { formatDuration, outputJson } from '@dd/core/helpers';
+import { outputJson } from '@dd/core/helpers/fs';
+import { formatDuration } from '@dd/core/helpers';
 import type { Logger } from '@dd/core/types';
 import type { MetricToSend, OutputOptions, Report } from '@dd/telemetry-plugin/types';
 import path from 'path';
