+++ packages/tools/src/commands/integrity/files.ts
@@ -2,7 +2,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { outputJsonSync, readJsonSync } from '@dd/core/helpers';
+import { outputJsonSync, readJsonSync } from '@dd/core/helpers/fs';
 import {
     CONFIGS_KEY,
     HELPERS_KEY,
